+++ src/main/java/com/project/delivery/Application.java
@@ -1,8 +1,10 @@
-package com.project.spring;
+package com.project.delivery;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 
+@EnableJpaAuditing
 @SpringBootApplication
 public class Application {
