+++ src/test/java/com/project/delivery/ApplicationTests.java
@@ -1,4 +1,4 @@
-package com.project.spring;
+package com.project.delivery;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
